[Feature][Connector-V2][Paimon] Rename flink-table-store to Paimon and using the 0.4-SNAPSHOT API
