--- seatunnel-connectors-v2/connector-flink-table-store/src/main/java/org/apache/seatunnel/connectors/seatunnel/fts/source/FlinkTableStoreSourceSplit.java
@@ -1,41 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.seatunnel.connectors.seatunnel.fts.source;
-
-import org.apache.seatunnel.api.source.SourceSplit;
-
-import org.apache.flink.table.store.table.source.Split;
-
-public class FlinkTableStoreSourceSplit implements SourceSplit {
-    private static final long serialVersionUID = 1L;
-
-    private final Split split;
-
-    public FlinkTableStoreSourceSplit(Split split) {
-        this.split = split;
-    }
-
-    @Override
-    public String splitId() {
-        return split.toString();
-    }
-
-    public Split getSplit() {
-        return split;
-    }
-}
